댓글 기능 수정, 좋아요기능 추가

+++ be/movies/views.py
@@ -7,6 +7,7 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated
 from .models import Movie, Comment
 from accounts.models import User
+from django.http import JsonResponse
 
 
 @api_view(['GET'])
@@ -34,22 +35,13 @@ def movie_detail(request, movie_pk):
             return Response(data, status=status.HTTP_204_NO_CONTENT)
 
 @api_view(['POST'])
-@permission_classes(['IsAuthenticated'])
-def comment_create(request, movie_pk):
-    print("request@@@@@@@@@@@@@@@@@")
-    print(request)
-    print(1)
+# @permission_classes(['IsAuthenticated'])
+def comment_create(request, movie_pk, username):
     movie = get_object_or_404(Movie, pk=movie_pk)
-    # user = get_object_or_404(User, pk=request.POST.get('user_id'))
+    userid = User.objects.get(username=username)
     serializer = CommentSerializer(data=request.data)
-    print(request.data)
-    print(1)
-    print(serializer)
     if serializer.is_valid(raise_exception=True):
-        print(5)
-        serializer.save(movie_id=movie.pk, user_id=1)
-        print(serializer.data)
-        print(4)
+        serializer.save(movie_id=movie.pk, user_id=userid.id)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
     else:
         print(3)
@@ -57,8 +49,9 @@ def comment_create(request, movie_pk):
 
 @api_view(['PUT', 'DELETE'])
 # @permission_classes(['IsAuthenticated'])
-def comment_detail(request, comment_pk):
+def comment_detail(request, comment_pk, username):
     comment = get_object_or_404(Comment, pk=comment_pk)
+    userid = User.objects.get(username=username)
     if request.method == 'PUT':
         # if request.user == comment.user:
             serializer = CommentCreateSerializer(comment, data=request.data)
@@ -73,15 +66,32 @@ def comment_detail(request, comment_pk):
         #     }    
         #     return Response(data, status=status.HTTP_401_UNAUTHORIZED)
     elif request.method == 'DELETE':
-        # if request.user == comment.user:
+        if userid.id == comment.user_id:
             comment.delete()
             data = {
                 'delete': f'comment {comment_pk} is deleted'
             }
             return Response(data, status=status.HTTP_204_NO_CONTENT)
-        # else:
-        #     data = {
-        #         'delete':False,
-        #         'description': '로그인한 유저가 작성한 글이 아닙니다.'
-        #     }    
-        #     return Response(data, status=status.HTTP_401_UNAUTHORIZED)
+        else:
+            data = {
+                'delete':False,
+                'description': '로그인한 유저가 작성한 글이 아닙니다.'
+            }    
+            return Response(data, status=status.HTTP_401_UNAUTHORIZED)
+
+@api_view(['POST'])
+def movie_like(request, movie_pk, username):
+    response = {'islike':''}
+    userid = User.objects.get(username=username)
+    if request.method == 'POST':
+        movie = get_object_or_404(Movie, pk=movie_pk)
+        existed = movie.like.filter(pk=userid.id).exists()
+        if existed:
+            movie.like.remove(userid.id)
+            response['islike'] = 'dislike'
+        else:
+            movie.like.add(userid.id)
+            response['islike'] = 'like'
+        return Response(response, status=status.HTTP_200_OK)
+    else:
+        return Response(response, status=status.HTTP_403_FORBIDDEN)
